Multivariate Outlier Detection

Identifies multivariate outliers using four different methods.

Usage

multiOutliers(
  data,
  varlist = names(data),
  method,
  minPts = 10,
  k = 5,
  threshold = 0.95,
  alpha = 0.1,
  ntrees = 100,
  n = 5,
  na.rm = TRUE,
  ...
)

Arguments

data

a data frame

varlist

a list of numeric variables

method

character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest

minPts

(optional) numeric, minimum points used for LoF outlier detection. Default value is 5

k

(optional) a k value used for the kNN method of outlier detection. Default value is 5

threshold

(optional) the threshold used for kNN outlier detection. Default value is 0.95

alpha

(optional) the alpha used for mahalanobis distance outlier detection. Default value is 0.1

ntrees

(optional) the number of trees used in iForest outlier detection. Default value is 100

n

(optional) the number of points to take as outliers in iForest outlier detection. Default value is 5

na.rm

(optional) logical, specifies whether to remove NA values. Defaults to TRUE

Value

method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters

Examples

multiOutliers(mtcarsOutliers, method="mahalanobis", alpha=0.1)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: mahalanobis
#> Dataset: mtcarsOutliers
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb scores
#> Row: 5 7 9 10 12 20 31 32
#> Outlier Score: 25.11784 29.58581 27.30969 19.22256 26.99406 21.46034 27.95683 25.39721
#> Message:  Outliers detected
#> Option 1 : alpha = 0.1
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> First Five Rows of Data Used:
#>                    mpg       cyl disp      hp drat       wt  qsec vs am gear
#> Mazda RX4         21.0  9.959246  160 110.000 3.90 2.620000 16.46  0  1    4
#> Mazda RX4 Wag     21.0  6.000000  160 110.000 3.90 2.875000 17.02  0  1    4
#> Datsun 710        22.8  4.000000  108  93.000 3.85 3.992371 18.61  1  1    4
#> Hornet 4 Drive    21.4  6.000000  258 110.000 3.08 3.215000 19.44  1  0    3
#> Hornet Sportabout 18.7 -4.137792  360 446.914 3.15 3.440000 17.02  0  0    3
#>                   carb    scores
#> Mazda RX4            4  7.247443
#> Mazda RX4 Wag        4  3.943746
#> Datsun 710           1  5.886045
#> Hornet 4 Drive       1  2.114513
#> Hornet Sportabout    2 25.117839
multiOutliers(mtcarsOutliers, method="LoF", minPts=5)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: LoF
#> Dataset: mtcarsOutliers
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb scores
#> Row: 5 7 9 10 11 12 19 20 29 31 32
#> Outlier Score: 3.146901 4.646142 3.401508 1.759811 2.00213 2.543073 1.915247 1.761172 2.28657 4.403093 1.772691
#> Message:  Outliers detected
#> Option 1 : minPts = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> First Five Rows of Data Used:
#>                    mpg       cyl disp      hp drat       wt  qsec vs am gear
#> Mazda RX4         21.0  9.959246  160 110.000 3.90 2.620000 16.46  0  1    4
#> Mazda RX4 Wag     21.0  6.000000  160 110.000 3.90 2.875000 17.02  0  1    4
#> Datsun 710        22.8  4.000000  108  93.000 3.85 3.992371 18.61  1  1    4
#> Hornet 4 Drive    21.4  6.000000  258 110.000 3.08 3.215000 19.44  1  0    3
#> Hornet Sportabout 18.7 -4.137792  360 446.914 3.15 3.440000 17.02  0  0    3
#>                   carb    scores
#> Mazda RX4            4 1.1569889
#> Mazda RX4 Wag        4 1.0824819
#> Datsun 710           1 0.9169562
#> Hornet 4 Drive       1 1.1209301
#> Hornet Sportabout    2 3.1469012
multiOutliers(mtcarsOutliers, method="kNN", k=5, threshold=.95)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: kNN
#> Dataset:
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb scores
#> Row: 11 19
#> Outlier Score: 547.8248 393.1049
#> Message:  Outliers detected
#> Option 1 : k = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> First Five Rows of Data Used:
#> [1] "mtcarsOutliers"
multiOutliers(mtcarsOutliers, method="iForest", ntrees = 50)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: iForest
#> Dataset: mtcarsOutliers
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb scores
#> Row: 5 9 10 20 31
#> Outlier Score: 0.6098171 0.5689141 0.5168466 0.4768288 0.4669088
#> Message:  Outliers detected
#> Option 1 : ntrees = 50
#> Option 2 : n = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> First Five Rows of Data Used:
#>                    mpg       cyl disp      hp drat       wt  qsec vs am gear
#> Mazda RX4         21.0  9.959246  160 110.000 3.90 2.620000 16.46  0  1    4
#> Mazda RX4 Wag     21.0  6.000000  160 110.000 3.90 2.875000 17.02  0  1    4
#> Datsun 710        22.8  4.000000  108  93.000 3.85 3.992371 18.61  1  1    4
#> Hornet 4 Drive    21.4  6.000000  258 110.000 3.08 3.215000 19.44  1  0    3
#> Hornet Sportabout 18.7 -4.137792  360 446.914 3.15 3.440000 17.02  0  0    3
#>                   carb    scores
#> Mazda RX4            4 0.4174958
#> Mazda RX4 Wag        4 0.3701169
#> Datsun 710           1 0.3681721
#> Hornet 4 Drive       1 0.3560603
#> Hornet Sportabout    2 0.4669088