Plotting Multivariate Outlier Detection

Plots identified multivariate outliers using either kNN, iForest, mahalanobis, LoF

Usage

# S3 method for class 'multiOutliers'
plot(
  data,
  varlist = names(data),
  methods = c("kNN", "LoF", "mahalanobis", "iForest"),
  minPts = 10,
  k = 5,
  threshold = 0.95,
  alpha = 0.1,
  na.rm = TRUE,
  ...
)

Arguments

x

the results from a multivariate outlier detection function (e.g., kNN, iForest, mahalanobis, LoF)

Value

A formatted plot of the results of the multivariate outlier detection

Examples

data(iris)
results <- multiOutliers(iris, method="iForest")
plot(results)
#> Error in data[, varlist]: incorrect number of dimensions