Printing Univariate Outlier Detection

Prints identified univariate outliers using boxplot, Grubbs test, or MAD method

Usage

# S3 method for class 'univOutliers'
print(x, ...)

Arguments

x

the results from a univariate outlier detection function (e.g., boxplot, grubbsTest, or MAD)

Value

A formatted print of the results of the univariate outlier detection

Examples

# Example usage:
object <- univOutliers(data = mtcars, method = "mad")
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
#> Error in plot.univOutliers(object, mtcars): could not find function "plot.univOutliers"
print(object)  # Printing the outliers and methods
#> Loading required package: cli
#> Loading required package: knitr
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: mpg
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: cyl
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: disp
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: hp
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: drat
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: wt
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: qsec
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      9             22.9      
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: vs
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      3               1       
#>      4               1       
#>      6               1       
#>      8               1       
#>      9               1       
#>      10              1       
#>      11              1       
#>      18              1       
#>      19              1       
#>      20              1       
#>      21              1       
#>      26              1       
#>      28              1       
#>      32              1       
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: am
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      1               1       
#>      2               1       
#>      3               1       
#>      18              1       
#>      19              1       
#>      20              1       
#>      26              1       
#>      27              1       
#>      28              1       
#>      29              1       
#>      30              1       
#>      31              1       
#>      32              1       
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: gear
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: mad ──────────────────────────────────────────────────────────
#> Variable: carb
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      31              8       

object <- univOutliers(data = mtcars, method = "boxplot")
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
#> Error in plot.univOutliers(object, mtcars): could not find function "plot.univOutliers"
print(object)  # Printing the outliers and methods
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: mpg
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: cyl
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: disp
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: hp
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      31             335      
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: drat
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: wt
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      16            5.424     
#>      17            5.345     
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: qsec
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      9             22.9      
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: vs
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: am
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: gear
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: boxplot ──────────────────────────────────────────────────────
#> Variable: carb
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      31              8       

object <- univOutliers(data = mtcars, method = "grubbs")
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
#> Warning: Data is not normally distributed. Grubbs' test may not be appropriate.
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
#> Error in plot.univOutliers(object, mtcars): could not find function "plot.univOutliers"
print(object)  # Printing the outliers and methods
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: mpg
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: cyl
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: disp
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: hp
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: drat
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: wt
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: qsec
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: vs
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: am
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: gear
#> ✔ No outliers detected.
#> 
#> ── Method Chosen: grubbs ───────────────────────────────────────────────────────
#> Variable: carb
#> 
#> Outliers Detected for:
#> 
#> 
#>  Row Number    Outlier Value 
#> ------------  ---------------
#>      31              8