qacOutliers: Identification of Multivariate and Univariate Outliers

The goal of qacOutliers is to detect multivariate and univariate outliers using seven different methods. This package consolidates methods from multiple different R packages and makes them more easily accessible.

See also

Author

Maintainer: Caleb Henning chenning@wesleyan.edu

Authors:

Larissa Xu lxu01@wesleyan.edu

Angelica Crown acrown@wesleyan.edu

Ernie Little elittle@wesleyan.edu

Braeden Falzarano bfalzarano@wesleyan.edu

Ral Reyes rreyes@wesleyan.edu

Tegh Singh tsingh@wesleyan.edu