Multivariate Outlier Detection

Identifies multivariate outliers using four different methods.

Usage

multiOutliers(
  data,
  varlist = names(data),
  method = "lof",
  minPts = 10,
  k = 5,
  threshold = 0.95,
  alpha = 0.1,
  ntrees = 100,
  n = 5,
  na.rm = TRUE
)

Arguments

data

a data frame

varlist

a list of variables. The variables can be factor or numeric. Categorical variables are dropped when using the mahalanobis and kNN methods.

method

character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest.

minPts

(optional) numeric, minimum points used for LoF outlier detection. Default value is 5

k

(optional) a k value used for the kNN method of outlier detection. Default value is 5

threshold

(optional) the threshold used for kNN outlier detection. Default value is 0.95

alpha

(optional) the alpha used for mahalanobis distance outlier detection. Default value is 0.1

ntrees

(optional) the number of trees used in iForest outlier detection. Default value is 100

n

(optional) the number of points to take as outliers in iForest outlier detection. Default value is 5

na.rm

(optional) logical, specifies whether to remove NA values. Defaults to TRUE. If the value is TRUE, removes missing data through listwise deletion.

Value

method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters

Examples

multiOutliers(mtcarsOutliers, method="mahalanobis", alpha=0.1)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: mahalanobis
#> Dataset: mtcarsOutliers
#> Variables: mpg disp hp drat wt qsec
#> Row: 3 4 9 16 19 20 24 30
#> Outlier Score: 21.23573 12.96992 17.48584 17.54522 11.54579 10.88054 11.19061 27.79833
#> Message:  Outliers detected
#> Option 1 : alpha = 0.1
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> Five Largest Outliers Within the Provided Dataset:
#>                          mpg      disp        hp      drat       wt      qsec
#> Ferrari Dino        19.70000  145.0000  175.0000 -5.146597  2.77000 15.500000
#> Datsun 710          22.80000 -468.2171   93.0000  3.850000 13.16772 18.610000
#> Lincoln Continental 10.40000  460.0000 -232.3693  3.000000  5.42400 17.820000
#> Merc 230            22.80000  140.8000   95.0000  3.920000  3.15000  3.655936
#> Hornet 4 Drive      54.27716  258.0000  110.0000  3.080000  3.21500 19.440000
#>                       scores
#> Ferrari Dino        27.79833
#> Datsun 710          21.23573
#> Lincoln Continental 17.54522
#> Merc 230            17.48584
#> Hornet 4 Drive      12.96992
multiOutliers(mtcarsOutliers, method="lof", minPts=5)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: LoF
#> Dataset: mtcarsOutliers
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb
#> Row: 1 2 3 4 6 7 8 9 10 12 13 15 17 20 21 23 27 28 29 30 31
#> Outlier Score: 1.082034 1.130627 1.010103 1.040641 1.055219 1.004195 1.015205 1.119456 1.005825 1.004323 1.003283 1.108191 1.078186 1.010103 1.067969 1.009432 1.229526 1.040732 1.87695 1.21083 1.461643
#> Message:  Outliers detected
#> Option 1 : minPts = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> Five Largest Outliers Within the Provided Dataset:
#>                     mpg cyl  disp  hp      drat    wt  qsec vs am gear carb
#> Ford Pantera L 15.80000   8 351.0 264  4.220000 3.170 14.50  0  1    5    4
#> Maserati Bora  15.00000   8 301.0 335  3.540000 3.570 14.60  0  1    5    8
#> Porsche 914-2  26.00000   4 120.3  91  4.430000 2.140 16.70  0  1    5    2
#> Ferrari Dino   19.70000   6 145.0 175 -5.146597 2.770 15.50  0  1    5    6
#> Mazda RX4 Wag  52.22202   6 160.0 110  3.555697 2.875 17.02  0  1    4    4
#>                  scores
#> Ford Pantera L 1.876950
#> Maserati Bora  1.461643
#> Porsche 914-2  1.229526
#> Ferrari Dino   1.210830
#> Mazda RX4 Wag  1.130627
multiOutliers(mtcarsOutliers, method="knn", k=5, threshold=.95)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: kNN
#> Dataset: mtcarsOutliers
#> Variables: mpg disp hp drat wt qsec
#> Row: 1 3 16
#> Outlier Score: 389.3906 557.6478 403.178
#> Message:  Outliers detected
#> Option 1 : k = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> Five Largest Outliers Within the Provided Dataset:
#>                      mpg      disp        hp drat       wt  qsec   scores
#> Datsun 710          22.8 -468.2171   93.0000 3.85 13.16772 18.61 557.6478
#> Lincoln Continental 10.4  460.0000 -232.3693 3.00  5.42400 17.82 403.1780
#> Mazda RX4           21.0  782.5020  110.0000 3.90  2.62000 16.46 389.3906
#> Honda Civic         30.4   75.7000 -240.5539 4.93  1.61500 18.52 312.8635
#> Maserati Bora       15.0  301.0000  335.0000 3.54  3.57000 14.60 122.3927
multiOutliers(mtcarsOutliers, method="iforest", ntrees = 50)
#> 
#> ── Summary Information ─────────────────────────────────────────────────────────
#> 
#> Method: iForest
#> Dataset: mtcarsOutliers
#> Variables: mpg cyl disp hp drat wt qsec vs am gear carb
#> Row: 2 3 4 9 30
#> Outlier Score: 0.5650616 0.5595005 0.5532602 0.5499185 0.5368699
#> Message:  Outliers detected
#> Option 1 : ntrees = 50
#> Option 2 : n = 5
#> 
#> ── Dataset Information ─────────────────────────────────────────────────────────
#> 
#> Five Largest Outliers Within the Provided Dataset:
#>                     mpg cyl      disp  hp      drat       wt      qsec vs am
#> Ferrari Dino   19.70000   6  145.0000 175 -5.146597  2.77000 15.500000  0  1
#> Datsun 710     22.80000   4 -468.2171  93  3.850000 13.16772 18.610000  1  1
#> Hornet 4 Drive 54.27716   6  258.0000 110  3.080000  3.21500 19.440000  1  0
#> Mazda RX4 Wag  52.22202   6  160.0000 110  3.555697  2.87500 17.020000  0  1
#> Merc 230       22.80000   4  140.8000  95  3.920000  3.15000  3.655936  1  0
#>                gear carb    scores
#> Ferrari Dino      5    6 0.5650616
#> Datsun 710        4    1 0.5595005
#> Hornet 4 Drive    3    1 0.5532602
#> Mazda RX4 Wag     4    4 0.5499185
#> Merc 230          4    2 0.5368699