Multivariate Outlier Detection

Identifies multivariate outliers using four different methods.

multiOutliers(
  data,
  varlist = names(data),
  method,
  minPts = 5,
  k = 5,
  threshold = 0.95,
  na.rm = TRUE,
  ...
)

Arguments

data

a data frame

varlist

a list of numeric variables

method

character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest

minPts

(optional) numeric, minimum points used for LoF outlier detection. Default value is 5

k

(optional) a k value used for the kNN method of outlier detection. Default value is 5

threshold

(optional) the threshold used for kNN outlier detection. Default value is 0.95

Value

indices of detected outliers, if any

Examples

data(mtcars)
multiOutliers(mtcars, method="mahalanobis")
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
#> $indices
#> named integer(0)
#> 
#> $values
#>  [1] X1  X2  X3  X4  X5  X6  X7  X8  X9  X10 X11
#> <0 rows> (or 0-length row.names)
#> 
multiOutliers(mtcars, method="LoF")
#>                     ID  mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4            1 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag        2 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710           3 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive       4 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout    5 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant              6 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 240D            8 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 280            10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Cadillac Fleetwood  15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Pontiac Firebird    25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Lotus Europa        28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#>                     LOF_Score
#> Mazda RX4            1.097199
#> Mazda RX4 Wag        1.097199
#> Datsun 710           1.012611
#> Hornet 4 Drive       1.251758
#> Hornet Sportabout    1.037223
#> Valiant              1.734751
#> Merc 240D            1.720491
#> Merc 280             1.097199
#> Merc 280C            1.097199
#> Cadillac Fleetwood   1.104841
#> Lincoln Continental  1.046855
#> Chrysler Imperial    1.046855
#> Fiat 128             1.195787
#> Honda Civic          1.195787
#> Toyota Corolla       1.195787
#> Toyota Corona        1.012057
#> Dodge Challenger     1.003272
#> AMC Javelin          1.031211
#> Pontiac Firebird     1.075088
#> Fiat X1-9            1.195787
#> Lotus Europa         1.214449
#> Ford Pantera L       1.238213
#> Ferrari Dino         1.909731
#> Maserati Bora        1.555531
multiOutliers(mtcars, method="kNN")
#> Loading required package: cli
#> Error in print.multiOutliers(x): object 'method' not found
multiOutliers(mtcars, method="iForest")
#> Due to small sample size, reduced 'min.node.size' to 11
#> 
#> Outlier identification by random forests
#> 
#>   Variables to check:		mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#>   Variables used to check:	mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#> 
#>   Checking: mpg  cyl  disp  hp  drat  wt  qsec  vs  am  gear  carb  
#>                 row  col observed  predicted      rmse    score threshold
#> Maserati.Bora    31   hp    335.0 207.542014 32.950648 3.868148         3
#> Merc.230          9 qsec     22.9  18.913611  1.119986 3.559320         3
#> Maserati.Bora.1  31 carb      8.0   4.007091  1.149726 3.472923         3
#>                 replacement
#> Maserati.Bora         335.0
#> Merc.230               22.9
#> Maserati.Bora.1         8.0